Move RegistrationFormFactory from user_api to user_authn.
And get_registration_extension_form.
