+++ common/djangoapps/student/forms.py
@@ -387,16 +387,3 @@ class AccountCreationForm(forms.Form):
             for key, value in self.cleaned_data.items()
             if key in self.extended_profile_fields and value is not None
         }
-
-
-def get_registration_extension_form(*args, **kwargs):
-    """
-    Convenience function for getting the custom form set in settings.REGISTRATION_EXTENSION_FORM.
-
-    An example form app for this can be found at http://github.com/open-craft/custom-form-app
-    """
-    if not getattr(settings, 'REGISTRATION_EXTENSION_FORM', None):
-        return None
-    module, klass = settings.REGISTRATION_EXTENSION_FORM.rsplit('.', 1)
-    module = import_module(module)
-    return getattr(module, klass)(*args, **kwargs)
